refactor: split api server and routes
